Enroute follow function(Mutual bottuns for driver and ad_client are done)
