+++ app/controllers/relationships_controller.rb
@@ -1,11 +1,12 @@
 class RelationshipsController < ApplicationController
+	# これは広告主→ドライバーをフォローするもの
 	def create
-		current_ad_client.follow(params[:ad_client_id])
+		current_ad_client.follow(params[:driver_id])
 		redirect_back(fallback_location: drivers_path)
 	end
 
 	def destroy
-		current_ad_client.unfollow(params[:ad_client_id])
+		current_ad_client.unfollow(params[:driver_id])
 		redirect_back(fallback_location: drivers_path)
 	end
 end
