--- app/helpers/relationship_ad_clients_helper.rb
@@ -1,2 +1,0 @@
-module RelationshipAdClientsHelper
-end
