+++ app/models/driver.rb
@@ -22,16 +22,16 @@ class Driver < ApplicationRecord
 	has_many :follower, class_name: "RelationshipDriver", foreign_key: "follower_id", dependent: :destroy # フォロー取得
 	has_many :followed, class_name: "RelationshipDriver", foreign_key: "followed_id", dependent: :destroy # フォロワー取得
 
-	has_many :following_driver, through: :follower, source: :followed #自分がフォローしている広告主
-	has_many :follower_driver, through: :followed, source: :follower #自分をフォローしている広告主
+	has_many :following_driver, through: :follower #source: :followed #自分がフォローしている広告主
+	has_many :follower_driver, through: :followed #source: :follower #自分をフォローしている広告主
 
-	 def follow(driver_id)
-	 	follower.create(followed_id: driver_id)
+	 def follow(ad_client_id)
+	 	follower.create(followed_id: ad_client_id)
 	 end
-	 def unfollow(driver_id)
-	 	follower.find_by(followed_id: driver_id).destroy
+	 def unfollow(ad_client_id)
+	 	follower.find_by(followed_id: ad_client_id).destroy
 	 end
-	 def following?(driver_id)
-	 	follower.find_by(followed_id: driver_id)
+	 def following?(ad_client_id)
+	 	follower.find_by(followed_id: ad_client_id)
 	 end
 end
