+++ app/views/ad_clients/show.html.erb
@@ -9,9 +9,9 @@
 				<%= attachment_image_tag @ad_client, :profile_image,:fill, 230, 170, fallback: "no_image.jpg", size:'230x170'  %><br>
 						<% if driver_signed_in? %> <!-- currentか引数を定義するか後で統一 -->
 	            <% if current_driver.following?(@ad_client) %>
-	              <%= link_to 'フォローを外す', driver_relationship_driver_path(@ad_client.id), method: :delete %>
+	              <%= link_to 'フォローを外す', ad_client_relationship_driver_path(@ad_client.id), method: :delete %>
 	            <% else %>
-	              <%= link_to 'フォローする', driver_relationship_drivers_path(@ad_client.id), method: :POST %>
+	              <%= link_to 'フォローする', ad_client_relationship_drivers_path(@ad_client.id), method: :POST %>
 	            <% end %>
 						<% end %>
 					<table class="table .table-borderless">
