+++ app/views/drivers/show.html.erb
@@ -3,15 +3,14 @@
 				<%= attachment_image_tag @driver, :profile_image,:fill, 230, 170, fallback: "no_image.jpg", size:'230x170'  %><br>
 						<% if ad_client_signed_in? %>
 	            <% if current_ad_client.following?(@driver) %>
-	              <%= link_to 'フォローを外す', ad_client_relationship_path(@driver.id), method: :delete %>
+	              <%= link_to 'フォローを外す', driver_relationship_path(@driver.id), method: :delete %>
 	            <% else %>
-	              <%= link_to 'フォローする', ad_client_relationships_path(@driver.id), method: :POST %>
+	              <%= link_to 'フォローする', driver_relationships_path(@driver.id), method: :POST %>
 	            <% end %>
 						<% end %>
 					<table class="table .table-borderless">
 						<tr>
-							<td>フォロー数：<%= link_to @driver.follower.count, followings_driver_path(@driver) %></td>
-      			  <td>フォロワー数：<%= link_to @driver.followed.count, followers_driver_path(@driver) %></td>
+
 						</tr>
 						<tr>
 							<th>名前</th>
