--- test/controllers/relationship_ad_clients_controller_test.rb
@@ -1,7 +1,0 @@
-require 'test_helper'
-
-class RelationshipAdClientsControllerTest < ActionDispatch::IntegrationTest
-  # test "the truth" do
-  #   assert true
-  # end
-end
